Completed validation, seekbar, animation and  reservation details functions
