+++ app/src/main/res/layout/activity_reservation_detail.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical">
@@ -163,7 +164,7 @@
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:layout_gravity="center"
-                        android:src="@drawable/ic_date_range_black_24dp" />
+                        app:srcCompat="@drawable/ic_date_range_black_24dp" />
 
                     <LinearLayout
                         android:id="@+id/ll_date"
@@ -177,7 +178,7 @@
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
                             android:layout_margin="8dp"
-                            android:text="@string/date"
+                            android:hint="@string/date"
                             android:textColor="@color/colorGrey"
                             android:textSize="16sp" />
 
@@ -193,7 +194,7 @@
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:layout_gravity="center"
-                        android:src="@drawable/ic_access_time_black_24dp" />
+                        app:srcCompat="@drawable/ic_access_time_black_24dp" />
 
                     <LinearLayout
                         android:id="@+id/ll_time"
@@ -207,7 +208,7 @@
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
                             android:layout_margin="8dp"
-                            android:text="@string/time"
+                            android:hint="@string/time"
                             android:textColor="@color/colorGrey"
                             android:textSize="16sp" />
 
@@ -237,19 +238,19 @@
                         android:textSize="16sp" />
 
                     <TextView
-                        android:id="@+id/tv_reserve_time"
+                        android:id="@+id/tv_reserve_selection_time"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:textColor="@color/colorAccent"
                         android:textSize="16sp" />
 
                     <TextView
-                        android:id="@+id/tv_reserve_remaining"
+                        android:id="@+id/tv_reserve_time"
                         android:layout_width="0dp"
                         android:layout_height="wrap_content"
                         android:layout_weight="1"
                         android:gravity="end"
-                        android:text=" -- min"
+                        android:text=""
                         android:textColor="@color/colorGrey"
                         android:textSize="16sp" />
