+++ app/src/main/res/menu/navigation.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/navigation_search"
+        android:icon="@drawable/ic_nav_search"
+        android:title="@string/search" />
+
+    <item
+        android:id="@+id/navigation_find"
+        android:icon="@drawable/ic_car"
+        android:title="@string/find" />
+
+    <item
+        android:id="@+id/navigation_reservation"
+        android:icon="@drawable/ic_reservation"
+        android:title="@string/reservation" />
+
+    <item
+        android:id="@+id/navigation_my_car"
+        android:icon="@drawable/ic_person_black_24dp"
+        android:title="@string/my_car" />
+
+</menu>
